% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_messages.R
\name{get_messages}
\alias{get_messages}
\title{Get messages from a channel.}
\usage{
get_messages(x, api_id, api_hash, n = 0, env = "r-telescrapeR")
}
\arguments{
\item{x}{Character vector containing the channel links.}

\item{api_id}{The API ID required for authentication.}

\item{api_hash}{The API hash required for authentication.}

\item{n}{Number of messages to be scraped. Defauls is 0, which means all.}

\item{env}{The name of the Python virtual environment to be used, when not "r-telescrapeR}
}
\value{
Returns a data frame.
}
\description{
This function sets the channel links, API ID, and API hash, and then retrieves messages from the specified channel using the provided API credentials.
}
