% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_telethon.R
\name{install_telethon}
\alias{install_telethon}
\title{Install the Telethon package in a specified Python virtual environment.}
\usage{
install_telethon(
  envname = "r-telescrapeR",
  new_env = identical(envname, "r-telescrapeR")
)
}
\arguments{
\item{envname}{The name of the Python virtual environment where Telethon will be installed. Default is "r-telescrapeR".}

\item{new_env}{Logical indicating whether a new virtual environment should be created. Default is \code{TRUE} if \code{envname} is "r-telescrapeR".}
}
\description{
This function installs the Telethon package, a Python library for Telegram API, in the specified Python virtual environment. If a new environment is requested and an environment with the same name already exists, it will be removed before installation.
}
\details{
The function installs the Telethon package in the specified Python virtual environment. If \code{new_env} is \code{TRUE} and the specified environment already exists, it will be removed before installation to ensure a clean setup.
}
\examples{
\dontrun{
# Install Telethon package in the default environment
install_telethon()

# Install Telethon package in a new environment named "myenv"
install_telethon(envname = "myenv", new_env = TRUE)

# Install Telethon package in an existing environment named "r-telescrapeR"
install_telethon(envname = "r-telescrapeR", new_env = FALSE)
}
}
