% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_messages.R
\name{start_ts}
\alias{start_ts}
\title{Setup the Telegram client}
\usage{
start_ts(api_id, api_hash, env = "r-telescrapeR")
}
\arguments{
\item{api_id}{The API ID provided by Telegram.}

\item{api_hash}{The API hash provided by Telegram.}

\item{env}{The name of the virtual environment to use. Default is "r-telescrapeR".}
}
\description{
This function loads the Python environment, the Python functions and defines the telegram client.
}
\examples{
\dontrun{
# Start TelescrapeR in default environment
start_ts(api_id, api_hash)
}
}
