% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_messages.R
\name{update_messages}
\alias{update_messages}
\title{Update message dataframe with new messages}
\usage{
update_messages(x, verbose = TRUE)
}
\arguments{
\item{x}{A data frame previously fetched by \code{telescrapeR}}

\item{verbose}{If TRUE, the function will print the channels and the number of messages being scraped.}
}
\value{
A data frame with old and new scraped messages.
}
\description{
\code{update_messages()} updates a data frame of previously fetched messages with newer messages. Use after \code{\link[=get_messages]{get_messages()}} to fetch new messages.
}
\examples{
\dontrun{
#Start TelescrapeR (if not already started)
start_ts(api_id, api_hash)

# Scrape all messages from channels
df <- get_messages(x)

# Update message dataframe with new messages
df_update <- update_messages(df)
}
}
\seealso{
\code{\link[=start_ts]{start_ts()}}, \code{\link[=get_messages]{get_messages()}}
}
