% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_messages.R
\name{get_messages}
\alias{get_messages}
\title{Get messages from a channel.}
\usage{
get_messages(x, n, api_id, api_hash, env = "r-telescraper")
}
\arguments{
\item{x}{Character vector containing the channel links.}

\item{n}{Number of messages to be scraped}

\item{api_id}{The API ID required for authentication.}

\item{api_hash}{The API hash required for authentication.}

\item{env}{The name of the Python virtual environment to be used, when not "r-telescrapeR}
}
\value{
A data frame with the following columns: channel_name, channel_id, title, message_id, message_views, date, sender_id, message_text
}
\description{
This function sets the channel links, API ID, and API hash, and then retrieves messages from the specified channel using the provided API credentials.
}
